+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.1 04/04/2022
 
-* TODO: Describe initial release.
+Initial Release supporting zoom inside sizedbox
+
+## 0.0.2 05/04/2022
+
+Supports zoom inside a scroll view